refactor: Extract into `FormatLinkableReferences`
