+++ app/services/autocomplete/autocomplete_linkable_references.rb
@@ -7,16 +7,7 @@ module Autocomplete
     end
 
     def call
-      search_results.map do |reference|
-        {
-          id: reference.id,
-          author: reference.author_names_string_with_suffix,
-          year: reference.citation_year,
-          title: reference.decorate.format_title,
-          full_pagination: full_pagination(reference),
-          bolton_key: bolton_key(reference)
-        }
-      end
+      Autocomplete::FormatLinkableReferences[search_results]
     end
 
     private
@@ -32,21 +23,6 @@ module Autocomplete
 
         match = Reference.find_by id: search_query
         [match] if match
-      end
-
-      def full_pagination reference
-        if reference.is_a? NestedReference
-          "[pagination: #{reference.pages_in} (#{reference.nesting_reference.pagination})]"
-        elsif reference.pagination
-          "[pagination: #{reference.pagination}]"
-        else
-          ""
-        end
-      end
-
-      def bolton_key reference
-        return "" unless reference.bolton_key
-        "[Bolton key: #{reference.bolton_key}]"
       end
   end
 end
